+++ tests/config/test_logging_actor.py
@@ -13,7 +13,7 @@ from frequenz.channels import Broadcast
 from marshmallow import ValidationError
 from pytest_mock import MockerFixture
 
-from frequenz.sdk.config import LoggerConfig, LoggingConfig, LoggingConfigUpdater
+from frequenz.sdk.config import LoggerConfig, LoggingConfig, LoggingConfigUpdatingActor
 
 
 def test_logging_config() -> None:
@@ -60,7 +60,7 @@ def cleanup_logs() -> Any:
     logging.getLogger("frequenz.sdk.timeseries").setLevel(logging.NOTSET)
 
 
-async def test_logging_config_updater_actor(
+async def test_logging_config_updating_actor(
     mocker: MockerFixture,
     cleanup_logs: Any,
 ) -> None:
@@ -70,11 +70,11 @@ async def test_logging_config_updater_actor(
     # Overriding logging.basicConfig would mess up other tests, so we mock it.
     # This is just for extra safety because changing root logging level in unit tests
     # is not working anyway - python ignores it.
-    mocker.patch("frequenz.sdk.config._logging_config_updater.logging.basicConfig")
+    mocker.patch("frequenz.sdk.config._logging_actor.logging.basicConfig")
 
     config_channel = Broadcast[Mapping[str, Any]](name="config")
     config_sender = config_channel.new_sender()
-    async with LoggingConfigUpdater(
+    async with LoggingConfigUpdatingActor(
         config_recv=config_channel.new_receiver().map(
             lambda app_config: app_config.get("logging", {})
         )